+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - refactor(core): make `get_eeprom_data` return a default value instead of `None` if there is an issue in procuring the data
 - refactor(docker): remove sync wrapper functions around async actions and make the action directly async as it is run by `subscribe_event` and it supports async functions, general housekeeping for docker service
 - fix: separate `has_output` and `has_output_stream` to fix the old `send_command` callers being broken with the new `send_command` implementation
+- feat: add a menu item in settings for installing/reinstalling the wm8960 driver
 
 ## Version 1.3.0